feat: refactor ad slideshow configuration and types for improved structure
