+++ src/renderer/components/states/announcement/AnnouncementContent.tsx
@@ -30,7 +30,8 @@ import useIslandStore from '../../../store/slices';
 import { AnnouncementHeader } from './components/AnnouncementHeader';
 import { AnnouncementBody } from './components/AnnouncementBody';
 import { ANNOUNCEMENT_DEFAULTS, ANNOUNCEMENT_KEYS } from './config/announcementDefaults';
-import { useAdSlides, AD_SLIDE_INTERVAL_MS } from './hooks/useAdSlides';
+import { useAdSlides } from './hooks/useAdSlides';
+import { AD_SLIDE_INTERVAL_MS } from './config/adSlidesDefaults';
 import { useAnnouncementData } from './hooks/useAnnouncementData';
 import { ProcessIndicator } from '../../components/DynamicIslandProcessIndicator';
 import { SvgIcon } from '../../../utils/SvgIcon';
