+++ src/renderer/components/states/announcement/test/AnnouncementContent.test.ts
@@ -34,23 +34,32 @@ const mocks = vi.hoisted(() => ({
   useAnnouncementData: vi.fn(),
 }));
 
-vi.mock('react', async (importOriginal) => ({
-  ...await importOriginal<typeof import('react')>(),
-  useState: (initialValue: unknown) => initialValue === true
-    ? [true, mocks.setListExpanded]
-    : [true, mocks.setShowVideo],
-}));
+vi.mock('react', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('react')>();
+  return {
+    ...actual,
+    useState: (initialValue: unknown) => initialValue === true
+      ? [true, mocks.setListExpanded]
+      : [true, mocks.setShowVideo],
+    useCallback: (fn: unknown) => fn,
+    useEffect: () => {},
+  };
+});
 vi.mock('react-i18next', () => ({
   useTranslation: () => ({
     t: (_key: string, options: string | { defaultValue: string }) =>
       typeof options === 'string' ? options : options.defaultValue,
   }),
+  initReactI18next: { type: '3rdParty', init: vi.fn() },
 }));
 vi.mock('../../../../store/slices', () => ({
   default: () => ({ setHover: vi.fn() }),
 }));
 vi.mock('../hooks/useAnnouncementData', () => ({
   useAnnouncementData: mocks.useAnnouncementData,
+}));
+vi.mock('../hooks/useAdSlides', () => ({
+  useAdSlides: () => [],
 }));
 
 import { AnnouncementContent } from '../AnnouncementContent';
@@ -86,8 +95,9 @@ describe('AnnouncementContent', () => {
     const panel = root.props.children as ReactElement<{ children: ReactElement }>;
     const detail = panel.props.children as ReactElement<{ children: ReactElement[] }>;
     const body = detail.props.children[2] as ReactElement<{ announcementList: ReactElement }>;
-    const list = body.props.announcementList as ReactElement<{ children: ReactElement<{ onClick: () => void }>[] }>;
-    const buttons = list.props.children;
+    const list = body.props.announcementList as ReactElement<{ children: ReactElement[] }>;
+    const nav = list.props.children[0] as ReactElement<{ children: ReactElement<{ onClick: () => void }>[] }>;
+    const buttons = nav.props.children;
     buttons[1].props.onClick();
 
     expect(buttons).toHaveLength(2);
